+++ src/main/java/com/nhlstenden/student/vigmo/services/LogService.java
@@ -27,7 +27,7 @@ public class LogService extends AbstractVigmoService<LogRepository, LogDto, Log>
     public long create(LogDto logDto) {
         //Will throw a data not found runtime exception if screen does not exist
         findUser(logDto.getUserId());
-        if(logDto.getDatetime() == null) logDto.setDatetime((double) Instant.now().getEpochSecond());
+        if(logDto.getDatetime() == null) logDto.setDatetime(Instant.now().getEpochSecond());
         return super.create(logDto);
     }
 
@@ -35,7 +35,7 @@ public class LogService extends AbstractVigmoService<LogRepository, LogDto, Log>
     public void update(LogDto logDto, long id) {
         //Will throw a data not found runtime exception if screen does not exist
         findUser(logDto.getUserId());
-        if(logDto.getDatetime() == null) logDto.setDatetime((double) Instant.now().getEpochSecond());
+        if(logDto.getDatetime() == null) logDto.setDatetime(Instant.now().getEpochSecond());
         super.update(logDto, id);
     }
